// lib/api/orval-generated/models/voucherState.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * The VoucherBook Complete API
 * Complete API documentation including public, admin, and internal endpoints
 * OpenAPI spec version: 1.0.0
 */

/**
 * Current state of the voucher lifecycle
 */
export type VoucherState = (typeof VoucherState)[keyof typeof VoucherState]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VoucherState = {
  draft: 'draft',
  published: 'published',
  claimed: 'claimed',
  redeemed: 'redeemed',
  expired: 'expired',
  suspended: 'suspended',
} as const