// lib/api/orval-generated/models/voucherDiscountType.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * The VoucherBook Complete API
 * Complete API documentation including public, admin, and internal endpoints
 * OpenAPI spec version: 1.0.0
 */

/**
 * Type of discount the voucher provides
 */
export type VoucherDiscountType =
  (typeof VoucherDiscountType)[keyof typeof VoucherDiscountType]

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const VoucherDiscountType = {
  fixed: 'fixed',
  percentage: 'percentage',
} as const